Revert "Merge branch 'main' of https://github.com/Em4nueel/connect-azul"
This reverts commit ca06ccfeadff79eca9e50dd3a5c4e1b1bd4282aa, reversing
changes made to 794c8dacbf83137d13c38cf8b6cd88b7c41aa849.

+++ src/main/java/com/connect/jpa/repository/UserInfoRepository.java
@@ -8,6 +8,6 @@ import com.connect.jpa.model.UserInfo;
 import java.util.Optional; 
   
 @Repository
-public interface UserInfoRepository extends JpaRepository<UserInfo, Long> { 
+public interface UserInfoRepository extends JpaRepository<UserInfo, Integer> { 
     Optional<UserInfo> findByName(String username); 
 }
